prometheus: Use output stream to return the metrics
Generating the metrics reply can take a long time, to reduce the load on
the system the prometheus use the output stream api.

This patch changes the message creation to use the output stream.
This both reduce the buffer size needed and make the prometheus yeild so
it will not block the shard.

The patch is not indented on purpose and an indentation patch will
follow.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>