metrics.cc: Add the labels when throwing duplicate metrics registration
The metrics layer throws an exception when registering the same metric
twice.

The metrics labels are part of the metrics uniqueness, so a metric id is
the same if it has the same name and the same label values.

When looking at a duplicate metric registration exception it will be
useful to see the label values as well.

This patch adds the labels to the name when throwing an exception.

The result will be:metrics_name,{label}={value}

For example:
std::runtime_error (registering metrics twice for metrics: io_queue_total_bytes,class=default,ioshard=0,mountpoint=none,shard=0,type=derive)

Signed-off-by: Amnon Heiman <amnon@scylladb.com>