metrics: Remove the type label
The metrics layer supports the concept of inherit_type which mean that
for reporting protocols that support it (collectd for example).

This mean that theoretically, in those protocols it possible to define
two metrics that differ only in their inherit type, for example
current_session and current_connection.

For protocols that do not support this, that type was added as a label.

In practice, it's an over-kill and just add information in the reporting
layer.

This patch remove the type label from the metrics, instead the
inherit_type will be placed in the metric_family_info.

Most of the changes are in scollectd, that now needs to get the
inherit_type from there.

Fixes #535